+++ CHANGELOG.md
@@ -73,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `X-Pxl-Cell` header renamed to `X-Pxl-Caller` for clarity.
 - Selected HTTP methods are now supposed to be provided in `open_api.selected_methods` instead of `methods`.
 - `headers` are now expected within `network` section.
+- Microsoft Excel add-in now auto-save services configuration as long as the modification are valid.
 
 ## [0.69.0] - 2018-12-03
 ### Changed